Make a Big Change. Try to increase the frequency of control. And find a program "MPU6050_DMP_CMP", which is a good example!

+++ FLYC/motor.h
@@ -1,2 +1,2 @@
 Servo motor[4];
-int Motor[4];
+uint16_t Motor[4];
